+++ sn_client/CHANGELOG.md
@@ -6,6 +6,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.99.20](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.19...sn_client-v0.99.20) - 2023-12-14
+
+### Other
+- updated the following local packages: sn_networking, sn_protocol, sn_registers, sn_transfers
+
+## [0.99.19](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.18...sn_client-v0.99.19) - 2023-12-14
+
+### Added
+- *(client)* add backoff to payment retries
+- *(networking)* use backoff for get_record
+
+## [0.99.18](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.17...sn_client-v0.99.18) - 2023-12-14
+
+### Other
+- *(test)* fix log messages during churn test
+
+## [0.99.17](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.16...sn_client-v0.99.17) - 2023-12-14
+
+### Added
+- *(cli)* simple retry mechanism for remaining chunks
+
+## [0.99.16](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.15...sn_client-v0.99.16) - 2023-12-13
+
+### Other
+- updated the following local packages: sn_networking
+
 ## [0.99.15](https://github.com/maidsafe/safe_network/compare/sn_client-v0.99.14...sn_client-v0.99.15) - 2023-12-13
 
 ### Added