+++ sn_faucet/CHANGELOG.md
@@ -6,6 +6,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.89](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.88...sn_faucet-v0.1.89) - 2023-12-14
+
+### Other
+- update dependencies
+
+## [0.1.88](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.87...sn_faucet-v0.1.88) - 2023-12-14
+
+### Other
+- update dependencies
+
+## [0.1.87](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.86...sn_faucet-v0.1.87) - 2023-12-14
+
+### Other
+- update dependencies
+
+## [0.1.86](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.85...sn_faucet-v0.1.86) - 2023-12-14
+
+### Other
+- update dependencies
+
+## [0.1.85](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.84...sn_faucet-v0.1.85) - 2023-12-14
+
+### Other
+- update dependencies
+
+## [0.1.84](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.83...sn_faucet-v0.1.84) - 2023-12-14
+
+### Other
+- update dependencies
+
+## [0.1.83](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.82...sn_faucet-v0.1.83) - 2023-12-13
+
+### Other
+- update dependencies
+
+## [0.1.82](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.81...sn_faucet-v0.1.82) - 2023-12-13
+
+### Other
+- update dependencies
+
 ## [0.1.81](https://github.com/maidsafe/safe_network/compare/sn_faucet-v0.1.80...sn_faucet-v0.1.81) - 2023-12-13
 
 ### Other